Refactor planeIndex to long
As all planar lengths are now longs, the rasterized plane index should
be a long as well. There could still be overflow issues with high plane
counts, but using a long index reduces that likelihood over an int.

+++ scifio/src/main/java/io/scif/AbstractWriter.java
@@ -102,7 +102,7 @@ public abstract class AbstractWriter<M extends TypedMetadata> extends
 	// -- Writer API Methods --
 
 	@Override
-	public void savePlane(final int imageIndex, final int planeIndex,
+	public void savePlane(final int imageIndex, final long planeIndex,
 		final Plane plane) throws FormatException, IOException
 	{
 		final long[] planeMax = metadata.getAxesLengthsPlanar(imageIndex);
@@ -312,7 +312,7 @@ public abstract class AbstractWriter<M extends TypedMetadata> extends
 	 * 
 	 * @throws FormatException if any of the arguments is invalid.
 	 */
-	protected void checkParams(final int imageIndex, final int planeIndex,
+	protected void checkParams(final int imageIndex, final long planeIndex,
 		final byte[] buf, final long[] planeMin, final long[] planeMax)
 		throws FormatException
 	{
