+++ scifio/src/main/java/io/scif/DefaultReader.java
@@ -69,7 +69,7 @@ public class DefaultReader extends ByteArrayReader<DefaultMetadata> implements
 	 * @throws UnsupportedOperationException
 	 */
 	@Override
-	public ByteArrayPlane openPlane(final int imageIndex, final int planeIndex,
+	public ByteArrayPlane openPlane(final int imageIndex, final long planeIndex,
 		final ByteArrayPlane plane, final long[] planeMin, final long[] planeMax)
 		throws FormatException, IOException
 	{
