+++ scifio/src/main/java/io/scif/TypedReader.java
@@ -62,11 +62,11 @@ public interface TypedReader<M extends TypedMetadata, P extends DataPlane<?>>
 {
 
 	@Override
-	P openPlane(int imageIndex, int planeIndex) throws FormatException,
+	P openPlane(int imageIndex, long planeIndex) throws FormatException,
 		IOException;
 
 	@Override
-	P openPlane(int imageIndex, int planeIndex, long[] planeMin, long[] planeMax)
+	P openPlane(int imageIndex, long planeIndex, long[] planeMin, long[] planeMax)
 		throws FormatException, IOException;
 
 	/**
@@ -76,7 +76,7 @@ public interface TypedReader<M extends TypedMetadata, P extends DataPlane<?>>
 	 * 
 	 * @see io.scif.Reader#openPlane(int, int, Plane)
 	 */
-	P openPlane(int imageIndex, int planeIndex, P plane) throws FormatException,
+	P openPlane(int imageIndex, long planeIndex, P plane) throws FormatException,
 		IOException;
 
 	/**
@@ -86,11 +86,11 @@ public interface TypedReader<M extends TypedMetadata, P extends DataPlane<?>>
 	 * 
 	 * @see io.scif.Reader#openPlane(int, int, Plane, long[], long[])
 	 */
-	P openPlane(int imageIndex, int planeIndex, P plane, long[] planeMin,
+	P openPlane(int imageIndex, long planeIndex, P plane, long[] planeMin,
 		long[] planeMax) throws FormatException, IOException;
 
 	@Override
-	P openThumbPlane(int imageIndex, int planeIndex) throws FormatException,
+	P openThumbPlane(int imageIndex, long planeIndex) throws FormatException,
 		IOException;
 
 	/**
